new_requirements.txt
amqp
antlr4-python3-runtime
arabic-reshaper
asgiref
asn1crypto
async-timeout
attrs
autobahn
Automat
billiard
bitarray
boto3
botocore
celery
certifi
cffi
channels
channels-redis
charset-normalizer
click
setuptools
wheel
click-didyoumean
click-plugins
click-repl
colorama
constantly
coreapi
coreschema
cron-descriptor
cryptography
cssselect2
Cython
daphne
Deprecated
dj-rest-auth
Django
django-cacheops
django-celery-beat
django-celery-results
django-cors-headers
django-extensions
django-oauth-toolkit
django-rest-swagger
django-ses
django-timezone-field
djangorestframework
djangorestframework-simplejwt
drf-yasg
exceptiongroup
filelock
funcy
gunicorn
html5lib
huggingface-hub
hydra-core
hyperlink
idna
incremental
inflection
iniconfig
install
itypes
Jinja2
jmespath
jwcrypto
kombu
llvmlite
lxml
MarkupSafe
more-itertools
mpmath
msgpack
networkx
oauthlib
omegaconf
triton
openai-whisper
openapi-codec
oscrypto
packaging
phonenumbers
Pillow
pluggy
portalocker
prompt-toolkit
psycopg2-binary
pyasn1
pyasn1-modules
pycparser
pyHanko
pyhanko-certvalidator
PyJWT
pyOpenSSL
pypdf
pypng
pytest
pytest-asyncio
pytest-django
python-bidi
python-crontab
python-dateutil
python-dotenv
python-flutterwave
pytz
pytz-deprecation-shim
PyYAML
qrcode
redis
regex
reportlab
requests
ruamel.yaml
ruamel.yaml.clib
s3transfer
sacrebleu
service-identity
simplejson
six
soundfile
sqlparse
svglib
sympy
tablib
tabulate
tiktoken
tinycss2
tomli
tqdm
transformers
Twisted
txaio
typing_extensions
tzdata
tzlocal
uritemplate
uritools
urllib3
vine
wcwidth
webencodings
whitenoise
wrapt
xhtml2pdf
zope.interface
